simple client start and stop test
